Implement page forward and backward.
Add keys to navigate a page forward (b) and backwards (B). A 'page' represents
the number of lines currently displayed in the terminal, which is approximated by the
terminal height. May be worthwhile in the future to find a more accurate way of
capturing the number of lines currently displayed.